chore: Override size method in LinkedHashMap (#687)
Motivation:
In sjsonnet, where I want to avoid array copy with pre-allocated
`Array`, which need to access the size of ujson's `LinkedHashMap`'s
`size()` method, by default, it will iterate the iterator to accumulate
the size.
But the underlying Java's `LinkedHashMap` has a dedicated field to get
the size, so we now avoid the `O(n)` iteration to a simple field read.

Modification:
Override the method `size` to make use of the underlying Java's
linkedHashMap's `size()` method.

Result:
Better performance.