Generalize `@flatten` annotation to work on any JSON-serializable key type rather than just `String` (#694)
This is necessary to allow the keys to be `BufferedValue`s or other
non-`String` data types, which is useful as it can store the `index:
Int` for error reporting purposes later

Some of the existing traits needed to be copy-pasted to allow
modification while leaving the old implementations in place for binary
compatibility.

Also bumped Mill to 1.1.0-RC3, moved the `mill-jvm-index` to `def jvmId`
configs, and bumped the various Scala/Scala.js/Scala-Native versions

Covered by some small unit tests. The Big150 test fails with a
stackoverflow on Scala 3.3.7, just commenting it out for now and hoping
that on more modern versions of Scala the compiler folks would have
fixed these pathological cases